refactor(api): Reorganize reference server `src/` by domain
Group multi-file subsystems into `storage/` (sqlite-store, file-store)
and `sync/` (doc-registry, websocket) directories, rename files to match
their exported class names, and rename `route-helpers.ts` to
`http-helpers.ts` to reflect its use from `sync/` as well as `routes/`.
Top-level `src/` is reduced from 19 files to 11.

No behavior change: `pnpm -F @v/api-spec generate` produces an identical
`openapi.yaml`, and all 62 tests still pass.
