+++ packages/api-server-reference/src/sync/websocket.ts
@@ -4,8 +4,8 @@ import * as decoding from 'lib0/decoding';
 import * as encoding from 'lib0/encoding';
 import * as syncProtocol from 'y-protocols/sync';
 
-import type { Deps } from './deps';
-import { toArrayBuffer } from './route-helpers';
+import type { Deps } from '../deps';
+import { toArrayBuffer } from '../http-helpers';
 
 const MESSAGE_SYNC = 0;
